Remove extra spaces

+++ main.py
@@ -4,11 +4,11 @@ from tqdm import trange
 
 
 def process_chunk(input, shift_amount):
-    data = np.fft.rfft(input)                                           # discrete fourier transform (find frequencies)
-    data = np.roll(data, -shift_amount)                                 # roll the array (remove shorter frequencies and pretend that remaining frequencies are them)
-    data[-shift_amount: len(data)] = 0                                  # add zeros to the end (shorter frequencies are rolled over to the end, let's remove them)
-    data = np.fft.irfft(data)                                           # inverse descrete fourier transform
-    data = data.astype(input.dtype)                                     # float array to int16
+    data = np.fft.rfft(input)             # discrete fourier transform (find frequencies)
+    data = np.roll(data, -shift_amount)   # roll the array (remove shorter frequencies and pretend that remaining frequencies are them)
+    data[-shift_amount: len(data)] = 0    # add zeros to the end (shorter frequencies are rolled over to the end, let's remove them)
+    data = np.fft.irfft(data)             # inverse descrete fourier transform
+    data = data.astype(input.dtype)       # float array to int16
     return data
